Pin pointers require fixed memory address in C code
