+++ src/lib.rs
@@ -2,12 +2,9 @@ mod lwip;
 mod mutex;
 mod output;
 mod stack;
-mod stack_impl;
 mod tcp_listener;
-mod tcp_listener_impl;
 mod tcp_stream;
 mod tcp_stream_context;
-mod tcp_stream_impl;
 mod udp;
 mod util;
